+++ src/registers/cp0_traits.rs
@@ -5,8 +5,8 @@ pub trait CP0Register {
     const REG_SEL : u8 = 0;
 }
 
-pub unsafe fn cp0_register_read<T: CP0Register>() -> usize {
-    let x: usize;
+pub unsafe fn cp0_register_read<T: CP0Register>() -> u32 {
+    let x: u32;
     asm!("mfc0 $0, $$$1, $2"
          : "=r"(x)
          : "i"(T::REG_ID), "i"(T::REG_SEL)
@@ -14,7 +14,7 @@ pub unsafe fn cp0_register_read<T: CP0Register>() -> usize {
     x
 }
 
-pub unsafe fn cp0_register_write<T: CP0Register>(val: usize) {
+pub unsafe fn cp0_register_write<T: CP0Register>(val: u32) {
     asm!("mtc0 $0, $$$1, $2"
          :
          : "r"(val), "i"(T::REG_ID), "i"(T::REG_SEL)
